Renamed legacy -> offset.
That code provides a generic parser for offside files into JSON.
Names 'LegacyToJSON' etc. will be reserved for the tools that build
upon the offset-code and provides usable LoKI-B input files in
JSON format.
